Refactor: Move default values to defalutValues file and apply it
