+++ src/hooks/super-admin/useSuperAdminWorkspace.tsx
@@ -1,5 +1,5 @@
 import { PaginationResponse, Workspace } from '@@types/index';
-import { defaultPaginationValue } from '@@types/PaginationType';
+import { defaultPaginationValue } from '@@types/defaultValues';
 import useApi from '@hooks/useApi';
 
 interface FetchAllWorkspacesParamsType {
