--- src/types/CategoryType.ts
@@ -1,9 +1,0 @@
-import { ProductCategory } from '.';
-
-export const defaultCategoryValue: ProductCategory = {
-  id: 0,
-  name: '',
-  index: null,
-  createdAt: '',
-  updatedAt: '',
-};
