--- src/types/ImageType.ts
@@ -1,8 +1,0 @@
-import { Image } from '.';
-
-export const defaultImageValue: Image = {
-  url: '',
-  id: 0,
-  createdAt: '',
-  updatedAt: '',
-};
