--- src/types/UserType.ts
@@ -1,11 +1,0 @@
-import { User, UserRole } from '.';
-
-export const defaultUserValue: User = {
-  name: '',
-  email: '',
-  role: UserRole.ADMIN,
-  accountUrl: '',
-  id: 0,
-  createdAt: '',
-  updatedAt: '',
-};
